refactor; move Verilog Token stuff to VerilogGlobals
