+++ VerilogToken/VerilogTokenTagger.cs
@@ -196,50 +196,12 @@ namespace VerilogLanguage.VerilogToken
             // 
             if (theNewText != theOldText)
             {
-                if (IsRefreshChar(theNewText) || IsRefreshChar(theOldText))
+                if (VerilogGlobals.IsRefreshChar(theNewText) || VerilogGlobals.IsRefreshChar(theOldText))
                 {
                     VerilogGlobals.Reparse(_buffer); // note that above, we are checking that the e.After is the same as the _buffer
                 }
             }
         }
-
-        // some characters, when encountered will need to have a full refresh, as they can have far-reaching consequences.
-        static public bool IsRefreshChar(string theString)
-        {
-            return (theString.Contains("/")) ||
-                   (theString.Contains("*")) ||
-                   (theString.Contains("[")) ||
-                   (theString.Contains("]")) ||
-                   (theString.Contains("}")) ||
-                   (theString.Contains("{")) ||
-                   (theString.Contains("(")) ||
-                   (theString.Contains(")"));  
-        }
-
-        static public bool IsDelimeter(string theString)
-        {
-            return (theString == " ") ||
-                   (theString == "~") ||
-                   (theString == "[") ||
-                   (theString == "]") ||
-                   (theString == "}") ||
-                   (theString == "{") ||
-                   (theString == "(") ||
-                   (theString == ")") ||
-                   (theString == ";") ||
-                   (theString == ",") ||
-                   (theString == "@") ||
-                   (theString == "\"") ||
-                   (theString == "\t");
-        }
-
-        static public bool IsEndingDelimeter(string theString)
-        {
-            return (theString == "]") ||
-                   (theString == "}") ||
-                   (theString == ")");
-        }
-
 
         //=====================================================================================
         // TODO - do spans always start on new lines? If not, we'll need an IsOpenLineComment
@@ -267,266 +229,9 @@ namespace VerilogLanguage.VerilogToken
             return isLocalBlockComment;
         }
 
-        public enum VerilogTokenContextType
-        {
-            Undetermined,
-            DoubleQuoteOpen,
-            SquareBracketOpen,
-            SquareBracketClose,
-            SquareBracketContents,
-            RoundBracketOpen,
-            RoundBracketClose,
-            RoundBracketContents,
-            SquigglyBracketOpen,
-            SquigglyBracketClose,
-            SquigglyBracketContents,
-            AlwaysAt,
-            Comment,
-            Text
-        }
 
 
-        private static VerilogTokenContextType VerilogTokenContextFromString(string s)
-        {
-            switch (s)
-            {
-                case null:
-                    return VerilogTokenContextType.Undetermined;
 
-                case "":
-                    return VerilogTokenContextType.Text;
-
-                default:
-                    switch (s.Substring(0, 1))
-                    {
-                        case "[":
-                            return VerilogTokenContextType.SquareBracketOpen;
-
-                        case "]":
-                            return VerilogTokenContextType.SquareBracketClose;
-
-                        case "(":
-                            return VerilogTokenContextType.RoundBracketOpen;
-
-                        case ")":
-                            return VerilogTokenContextType.RoundBracketClose;
-
-                        case "{":
-                            return VerilogTokenContextType.SquigglyBracketOpen;
-
-                        case "}":
-                            return VerilogTokenContextType.SquigglyBracketClose;
-
-                        case "@":
-                            return VerilogTokenContextType.AlwaysAt;
-
-                        default:
-                            return VerilogTokenContextType.Text;
-
-                    }
-            }
-        }
-
-        /// <summary>
-        ///   VerilogParseState - while processing each segment, we'll keep track of attributes in a VerilogParseState 
-        /// </summary>
-        public struct VerilogParseState
-        {
-            public string thisItem;
-            public int thisIndex;
-            public string priorChar;
-            public string priorDelimiter;
-            public bool thisCharIsDelimiter;
-            public bool priorCharIsDelimiter;
-
-            public bool thisCharIsEndingDelimiter;
-            public bool priorCharIsIsEndingDelimiter;
-
-            public bool hasOpenSquareBracket;
-            public bool hasOpenRoundBracket;
-            public bool hasOpenSquigglyBracket;
-
-            public bool hasOpenDoubleQuote;
-
-            public bool IsNewDelimitedSegment;
-
-            private string _thisChar;
-            public string thisChar
-            {
-                get { return _thisChar; }
-                set
-                {
-                    _thisChar = value;
-                    thisCharIsDelimiter = IsDelimeter(value);
-                    thisCharIsEndingDelimiter = IsEndingDelimeter(value);
-                    priorCharIsDelimiter = IsDelimeter(priorChar);
-                    // note  contiguous spaces are a single segment
-                    IsNewDelimitedSegment = ( thisCharIsDelimiter || priorCharIsDelimiter) && !((_thisChar == " ") && (priorChar == " "));
-
-                    if (IsNewDelimitedSegment)
-                    {
-
-                    }
-                    else
-                    {
-                        thisItem += thisChar;
-                    }
-                }
-            }
-
-            public void SetPriorValues()
-            {
-                // return;
-                priorCharIsDelimiter = thisCharIsDelimiter;
-                priorCharIsIsEndingDelimiter = thisCharIsEndingDelimiter;
-
-                priorChar = thisChar;
-                if (thisCharIsDelimiter)
-                {
-                    priorDelimiter = thisChar;
-                }
-            }
-
-
-            // initialize this VerilogParseState at creation time
-            public VerilogParseState(int i)
-            {
-                _thisChar = "";
-                thisIndex = 0;
-                thisItem = "";
-                IsNewDelimitedSegment = false;
-                priorChar = "";
-                priorDelimiter = "";
-                thisCharIsDelimiter = false;
-                priorCharIsDelimiter = false;
-                thisCharIsEndingDelimiter = false;
-                priorCharIsIsEndingDelimiter = false;
-                hasOpenSquareBracket = false;
-                hasOpenRoundBracket = false;
-                hasOpenSquigglyBracket = false;
-                hasOpenDoubleQuote = false;
-            }
-        }
-
-
-        /// <summary>
-        ///   VerilogToken
-        /// </summary>
-        public struct VerilogToken
-        {
-            public VerilogParseState ParseState;
-            public string Part;
-            public VerilogTokenContextType Context;
-
-            public void SetContext()
-            {
-                if (ParseState.hasOpenDoubleQuote)
-                {
-                    Context = VerilogTokenContextType.DoubleQuoteOpen;
-                }
-                else
-                {
-                    if (ParseState.hasOpenSquareBracket && !IsDelimeter(ParseState.thisChar))
-                    {
-                        Context = VerilogTokenContextType.SquareBracketContents;
-                    }
-                    else
-                    {
-                        if (ParseState.hasOpenRoundBracket && !IsDelimeter(ParseState.thisChar))
-                        {
-                            Context = VerilogTokenContextType.RoundBracketContents;
-                        }
-                        else
-                        {
-                            if (ParseState.hasOpenRoundBracket && !IsDelimeter(ParseState.thisChar))
-                            {
-                                Context = VerilogTokenContextType.SquigglyBracketContents;
-                            }
-                            else
-                            {
-                                Context = VerilogTokenContextFromString(ParseState.thisChar);
-                            }
-                        }
-                    }
-                }
-
-
-            }
-
-
-            /// <summary>
-            ///   Verilog Token Initializer
-            /// </summary>
-            /// <param name="p"></param>
-            /// <param name="c"></param>
-            public VerilogToken(string p = "", VerilogTokenContextType c = VerilogTokenContextType.Undetermined)
-            {
-                ParseState = new VerilogParseState(0);
-                Part = p ?? ""; // ensure Part is never null (empty string if p is null)
-
-                if (c == VerilogTokenContextType.Undetermined && Part.Length > 0)
-                {
-                    Context = VerilogTokenContextFromString(p); // we'll figure out the context from the first character
-                }
-                else
-                {
-                    Context = c; // unless otherwise specified
-                }
-            }
-        }
-
-        /// <summary>
-        ///    VerilogToken[] 
-        /// </summary>
-        /// <param name="theString"></param>
-        /// <returns></returns>
-        public VerilogToken[] VerilogKeywordSplit(string theString, VerilogToken priorToken)
-        {
-            List<VerilogToken> tokens = new List<VerilogToken>();
-            VerilogToken thisToken = new VerilogToken();
-            VerilogParseState thisContinuedParseState = new VerilogParseState(0);
-
-            // AddToken - appends the current token part to the array and create a new thisToken to build
-            void AddToken()
-            {
-                string thisItem = thisToken.ParseState.thisItem;
-                if (thisItem != "")
-                {
-                    thisToken.Part = thisToken.ParseState.thisItem;
-                    thisContinuedParseState = thisToken.ParseState;
-                    tokens.Add(thisToken);
-                    thisToken = new VerilogToken(thisToken.ParseState.thisChar);
-                    thisToken.ParseState = thisContinuedParseState;
-                    thisToken.ParseState.thisItem = thisToken.ParseState.thisChar; // start building a new token with the current, non-delimiter character, will be used to determine context in VerilogTokenContextFromString
-                }
-            }
-
-            thisToken.ParseState = priorToken.ParseState; // when starting, use the priorToken parseState that wouldhave come from the prior line in the span
-
-            for (int i = 0; i < theString.Length; i++)
-            {
-                thisToken.ParseState.thisIndex = i;
-                thisToken.ParseState.thisChar = theString.Substring(i, 1); // note setting this values triggers ParseState attribute assignments
-
-                if (thisToken.ParseState.IsNewDelimitedSegment)
-                {
-                    // anytime a delimiter is encountered, we start a new text segment 
-                    // note the delimiter itself is in a colorizable segment
-
-                    // there's a new delimiter, so add the current item and prep for the next one
-                    AddToken();
-
-                    // once the ParseState is configured (above, when assigning thisChar), set the context of the item
-                    thisToken.SetContext(); // TOFO do we really need this? context is alreasy set
-                    // at the end of each loop, set the prior values
-                    thisToken.ParseState.SetPriorValues();
-                } // end of for loop look at each char
-            }
-
-            // if there's anythin left, add it is as token (blank token not added)
-            AddToken();
-            return tokens.ToArray();
-        }
 
         /// <summary>
         ///   IEnumerable VerilogTokenTag GetTags
@@ -541,8 +246,8 @@ namespace VerilogLanguage.VerilogToken
             // init
             VerilogGlobals.InitHoverBuilder();
             VerilogGlobals.IsContinuedBlockComment = IsOpenBlockComment(spans); // TODO - does spans always contain the full document? (appears perhaps not)
-            VerilogToken[] tokens = null;
-            VerilogToken priorToken = new VerilogToken();
+            VerilogGlobals.VerilogToken[] tokens = null;
+            VerilogGlobals.VerilogToken priorToken = new VerilogGlobals.VerilogToken();
 
             // look at each span for tokens, comments, etc
             foreach (SnapshotSpan curSpan in spans)
@@ -556,10 +261,10 @@ namespace VerilogLanguage.VerilogToken
                 int LinePosition = 0;
                 string thisTokenString = "";
 
-                tokens = VerilogKeywordSplit(containingLine.GetText(), priorToken);
+                tokens = VerilogGlobals.VerilogKeywordSplit(containingLine.GetText(), priorToken);
 
                 Boolean IsContinuedLineComment = false; // comments with "//" are only effective for the current line, but /* can span multiple lines
-                foreach (VerilogToken VerilogToken in tokens) // this group of tokens in in a single line
+                foreach (VerilogGlobals.VerilogToken VerilogToken in tokens) // this group of tokens in in a single line
                 {
                     // by the time we get here, we might have a tag with adjacent comments:
                     //     assign//
@@ -624,8 +329,8 @@ namespace VerilogLanguage.VerilogToken
                                     // int testValue = VerilogGlobals.BracketDepth(containingLine.LineNumber, curLoc - containingLine.Start.Position);
                                     switch (VerilogToken.Context)
                                     {
-                                        case VerilogTokenContextType.SquareBracketOpen:
-                                        case VerilogTokenContextType.SquareBracketClose:
+                                        case VerilogGlobals.VerilogTokenContextType.SquareBracketOpen:
+                                        case VerilogGlobals.VerilogTokenContextType.SquareBracketClose:
                                             thisDelimiterTotalDepth = VerilogGlobals.BracketDepth(containingLine.LineNumber, curLoc - containingLine.Start.Position);
                                             thisDelimiterIndex = (thisDelimiterTotalDepth % 5);
                                             if (tokenSpan.IntersectsWith(curSpan))
@@ -634,8 +339,8 @@ namespace VerilogLanguage.VerilogToken
                                                                                       new VerilogTokenTag(_VerilogTypes["bracket_type" + (thisDelimiterIndex).ToString()]));
                                             break;
 
-                                        case VerilogTokenContextType.RoundBracketClose:
-                                        case VerilogTokenContextType.RoundBracketOpen:
+                                        case VerilogGlobals.VerilogTokenContextType.RoundBracketClose:
+                                        case VerilogGlobals.VerilogTokenContextType.RoundBracketOpen:
                                             thisDelimiterTotalDepth = VerilogGlobals.BracketDepth(containingLine.LineNumber, curLoc - containingLine.Start.Position);
                                             thisDelimiterIndex = (thisDelimiterTotalDepth % 5);
                                             if (tokenSpan.IntersectsWith(curSpan))
@@ -644,8 +349,8 @@ namespace VerilogLanguage.VerilogToken
                                                                                       new VerilogTokenTag(_VerilogTypes["bracket_type" + (thisDelimiterIndex).ToString()]));
                                             break;
 
-                                        case VerilogTokenContextType.SquigglyBracketOpen:
-                                        case VerilogTokenContextType.SquigglyBracketClose:
+                                        case VerilogGlobals.VerilogTokenContextType.SquigglyBracketOpen:
+                                        case VerilogGlobals.VerilogTokenContextType.SquigglyBracketClose:
                                             thisDelimiterTotalDepth = VerilogGlobals.BracketDepth(containingLine.LineNumber, curLoc - containingLine.Start.Position);
                                             thisDelimiterIndex = (thisDelimiterTotalDepth % 5);
                                             if (tokenSpan.IntersectsWith(curSpan))
@@ -654,13 +359,13 @@ namespace VerilogLanguage.VerilogToken
                                                                                       new VerilogTokenTag(_VerilogTypes["bracket_type" + (thisDelimiterIndex).ToString()]));
                                             break;
 
-                                        case VerilogTokenContextType.SquareBracketContents:
+                                        case VerilogGlobals.VerilogTokenContextType.SquareBracketContents:
                                             if (tokenSpan.IntersectsWith(curSpan))
                                                 yield return new TagSpan<VerilogTokenTag>(tokenSpan,
                                                                                       new VerilogTokenTag(VerilogTokenTypes.Verilog_BracketContent));
                                             break;
 
-                                        case VerilogTokenContextType.AlwaysAt:
+                                        case VerilogGlobals.VerilogTokenContextType.AlwaysAt:
                                             if (tokenSpan.IntersectsWith(curSpan))
                                                 yield return new TagSpan<VerilogTokenTag>(tokenSpan,
                                                                                       new VerilogTokenTag(VerilogTokenTypes.Verilog_always));
